feat: Place의 지도 조회 완성 KAN-103
