+++ frontend/src/components/place/place-item/PlaceItem.jsx
@@ -1,3 +1,4 @@
+import { useLocation } from 'react-router-dom';
 import {
   FooterDetails,
   ItemDetails,
@@ -10,16 +11,19 @@ import {
 } from '../place-liked/PlaceLiked.style';
 import { StarFilled } from '@ant-design/icons';
 import styled from 'styled-components';
-import CheckInactive from '/src/assets/icons/check-circle-inactive.png'
-import CheckActive from '/src/assets/icons/check-circle-active.png'
+import CheckInactive from '/src/assets/icons/check-circle-inactive.png';
+import CheckActive from '/src/assets/icons/check-circle-active.png';
 
-const CheckBox = styled.img`
-`
+const CheckBox = styled.img``;
 
 function PlaceItem({ isSelected, srcImg, name, star, category, info }) {
+  // 현재 URL
+  const location = useLocation();
+  const currentUrl = location.pathname;
+  
   return (
     <LikeContainer>
-      <LikeItem style={{paddingRight: '20px'}}>
+      <LikeItem style={{ paddingRight: '20px' }}>
         <ItemImage src={srcImg} alt="Five Guys" />
         <ItemDetails>
           <ItemTitle>{name}</ItemTitle>
@@ -29,7 +33,9 @@ function PlaceItem({ isSelected, srcImg, name, star, category, info }) {
           </ItemRating>
           <FooterDetails>{info}</FooterDetails>
         </ItemDetails>
-        <CheckBox style={{width: '32px'}} src={isSelected ? CheckActive : CheckInactive} />
+        {currentUrl == '/coursecreate' && (
+          <CheckBox style={{ width: '32px' }} src={isSelected ? CheckActive : CheckInactive} />
+        )}
       </LikeItem>
     </LikeContainer>
   );
